Add files via upload
Continuing with chapter 8. We were taught about learning rates of a model. You want to find the best learning rate: not too fast because the accuracy will be low), not too slow (because it will be underfitted... you could have learned more), but just right. Makes me think of Goldilocks.